Refactored out common code to remove inheritance
